refactor: split json details highlighter
